+++ targets/arm/mikroe/renesas/include/hal_ll_mstpcr.h
@@ -59,7 +59,8 @@ extern "C"{
     #define _MSTPCRC   ( uint32_t * )0x40047004
     #define _MSTPCRD   ( uint32_t * )0x40047008
 #elif (defined(R7FA4M2) || defined(R7FA4M3) || defined(R7FA6M4) || \
-       defined(R7FA6M5) || defined(R7FA4L1) || defined(R7FA6E2))
+       defined(R7FA6M5) || defined(R7FA6E2) || defined(R7FA4L1) || \
+       defined(R7FA4E2))
     #define _MSTPCRA   ( uint32_t * )0x40084000
     #define _MSTPCRB   ( uint32_t * )0x40084004
     #define _MSTPCRC   ( uint32_t * )0x40084008
@@ -124,7 +125,7 @@ typedef struct
 
     #if (defined(R7FA4M1) || defined(R7FA4M2) || defined(R7FA4M3) || \
          defined(R7FA6M3) || defined(R7FA6M4) || defined(R7FA6M5) || \
-         defined(R7FA4L1) || defined(R7FA6E2))
+         defined(R7FA4L1) || defined(R7FA6E2) || defined(R7FA4E2))
     uint32_t pclka;   // PCLKA clock frequency in Hz
     uint32_t pclkb;   // PCLKB clock frequency in Hz
     uint32_t pclkc;   // PCLKC clock frequency in Hz