+++ changelog/v2.17.12/changelog.md
@@ -16,7 +16,33 @@
 
 - [`v2.17.12`](#v21712)
   - [Changes](#changes)
+    - [RENESAS](#renesas)
     - [NEW HARDWARE](#new-hardware)
+
+### <font color=red>RENESAS</font>
+
+**RA4E2 MCUs supported!**
+
+|           |
+| :-----------: |
+| R7FA4E2B92CBB |
+| R7FA4E2B92CBC |
+| R7FA4E2B93CBB |
+| R7FA4E2B93CBC |
+| R7FA4E2B93CFM |
+| R7FA4E2B93CNE |
+| R7FA4E2B93CNH |
+|           |
+
+**Supported modules:**
+
++ ADC (Full module support)
++ SPI (Full module support)
++ UART (Full module support)
++ I2C (Full module support)
++ PWM (Full module support)
++ GPIO (Full module support)
++ 1-Wire (Full module support)
 
 ### NEW HARDWARE